[Backend] Clean up MMAv3 part of MatmulLoopPipeline.
The MMAv3 pipelining code needed some love and comments.

There are some minor functional changes in here.  For example, the old
code relied on a particular order of op.users(), and I'm not convinced
it was correct.

Once you unravel the code, what it needs to do is a lot less complicated
than what it was doing.  Indeed, despite the fact that I added a ton of
comments, it's roughly the same length as before.

The changes to the test are cosmetic.  Previously we emitted tt.dot and
assumed it would eventually becdome ttng.dot_async in a later pass (or
at least during codegen).  Now we explicitly convert some tt.dot's to
ttng.dot_async so we can have an explicit wait and thread the results
through.

There's more cleanup that can be done here (see the TODOs added in this
PR), but I needed to stop somewhere.

GPC: mmav3-cleanup